chg: test: rename imported test helpers to avoid any discovery issues
